implement /package-upload-url route in new API
